# s2: Slightly faster non-assembly decompression